refactor: port InsertSheetModal to runes mode
